add one more check

+++ second.sh
@@ -19,6 +19,12 @@ else
   nmtui
 fi
 
+if [ "$(whoami)" != "aneesh" ]; then
+        echo "Error: Please reboot and run the script as user aneesh."
+        echo "Not rebooting but instead logging out and logging in as aneesh will cause the script to break."
+        exit
+fi
+
 if [[ -f "/etc/paxxer-first-done" ]]; then
     echo "First.sh has been run, good."
     sudo rm -rf /etc/paxxer-first-done
